requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.4
django-admin-bootstrapped==2.5.7
django-authtools==1.6.0
django-braces==1.11.0
django-cors-middleware==1.3.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.4
django-extensions==1.9.0
django-filter==1.0.4
django-forms-bootstrap==3.1.0
django-oauth-toolkit==1.0.0
django-registration==2.2
django-registration-redux==1.7
djangorestframework==3.6.4
drfdocs==0.0.11
easy-thumbnails==2.4.1
idna==2.6
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
oauthlib==2.0.2
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
